fix `config set` bugs with auto field setting
